html/webappapis/scripting/events/event-handler-attributes-body-window.html

<!DOCTYPE html>
<title>HTMLBodyElement event handlers</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/WebIDLParser.js"></script>
<script src="resources/event-handler-body.js"></script>
<div id="log"></div>
<body>
<script>
setup({ explicit_done: true });

function f() {
  return 0;
}

handlersListPromise.then(({ shadowedHandlers }) => {
  shadowedHandlers.forEach(function(handler) {
    test(function() {
      document.body['on' + handler] = f;
      assert_equals(window['on' + handler], f);
    }, handler);
  });

  shadowedHandlers.forEach(function(handler) {
    document.body['on' + handler] = null;
  });

  shadowedHandlers.forEach(function(handler) {
    test(function() {
      assert_equals(document.body['on' + handler], null);
      assert_equals(window['on' + handler], null);
    }, handler + " removal");
  });

  done();
});
</script>
